refactor: split transaction tests into filtering, crud, and display files
